Move Exporter to /collector so it can be used as library
